refactor: move stats fetcher implementation to different files
