+++ cloud/storage/core/libs/diagnostics/ya.make
@@ -3,6 +3,7 @@ LIBRARY()
 SRCS(
     busy_idle_calculator.cpp
     counters_helper.cpp
+    cgroup_stats_fetcher.cpp
     critical_events.cpp
     executor_counters.cpp
     histogram_types.cpp
@@ -16,6 +17,7 @@ SRCS(
     solomon_counters.cpp
     stats_fetcher.cpp
     stats_updater.cpp
+    task_stats_fetcher.cpp
     trace_processor_mon.cpp
     trace_processor.cpp
     trace_reader.cpp
